Fix: Silence FlyGrep opens to avoid hit-enter-prompt
When using FlyGrep outside of SpaceVim, I get the "Press ENTER or type command to continue" prompt when opening a selection with FlyGrep. I couldn't make this go away with shortmess changes.

I validated this change in the standalone plugin and I followed the existing conventions of the use of 'silent' in the flygrep.vim file (silence the command rather than the 'exe').