Fix: enable clean startup in VimR
This is a follow-on to #1651; I found that I was still seeing the error message
mentioned in that PR under VimR 0.32.0.  Fortunately VimR has a specific
feature `gui_vimr` to test against; recent builds of VimR do set `gui_running`,
so the fix from that previous PR no longer works.