:bug: cache init.toml based on filename if linked (#3762)
This commit fixes the following edgecase:

If init.toml is a symlink to another file and that file timestamp is not
updated when the file is changed then the caching system breaks.

This happens in particular for users that rely on [NixOS](https://nixos.org/) or
[home-manager](https://github.com/rycee/home-manager) to
manage their vim configuration. This is because in Nix all files have
the same timestamp of the UNIX EPOCH. When generating a new
version of the file Nix will checksum the name (hence changing the name
the links point to) but the timestamp will ALWAYS be the same.