fetch_repo: Never change a directory
It is bad practice to change the currents shell working directory. E.g.
never use 'cd' in a script.

For example, if `git pull` crashes, it would leave the user in the git
directory 'somewhere' on his filesystem, potentially causing confusion,
as the following `cd -` command is never executed.

There are of course always exceptions, such as when tooling does not
(easily) support out of tree invocation.

To solve this, run the command in a subshell, which won't touch the
current working directory as the exit of the subshell kept the previous
working directory.

A more difficult alternative would be to pass `--git-dir` and
`--work-tree` along to git, but this would make it far harder to read.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>